requirements.txt
fastapi
uvicorn
pypdf2
python-docx
langchain
chromadb
transformers
openai
google-generativeai
tiktoken
spacy 